+++ frontend/src/component/feature/view-container.jsx
@@ -13,11 +13,6 @@ import { hasPermission } from '../../permissions';
 export default connect(
     (state, props) => ({
         features: state.features.toJS(),
-        betaFlags: state.features
-            .toJS()
-            .filter(t => t.enabled)
-            .filter(t => t.name.startsWith('unleash.beta'))
-            .map(t => t.name),
         featureToggle: state.features.toJS().find(toggle => toggle.name === props.featureToggleName),
         activeTab: props.activeTab,
         hasPermission: hasPermission.bind(null, state.user.get('profile')),